Pull out simplify() to optimize.py
This function has specific knowledge of Nand which is otherwise not needed to evaluate arbitrary chips.
